Add files via upload
We propose an in-vivo, untethered capsule equipped with three sensors, a wireless transmission module, and a locomotion mechanism. Doctors and physicians will be able to prescribe this mechanism for their patients. The main features of this microtechnology are to measure drug levels, blood glucose and oxygen levels, and blood pressure. The sensors on the device would perform continuous, real-time monitoring and would send a signal to a smart device to record the data. If the sensors detect excess quantities, they will send notifications/alerts to the doctor and patient via a mobile application that is connected to the device. Especially during COVID-19, health issues such as opioid overdoses are reaching an all-time high and our invention makes it so that patients’ conditions are monitored to prevent an overdose, heart attack, or stroke.